// frontend/src/app/models/UserModel.ts
export class UserModel{

    _id: any;
    username: String;
    password: String;
    email: String;
    name: String;
    surname: String;
    gender: Boolean;
    dateOfBirth: Date;
    profileImageURL: String;
    autQuestion: String;
    autAnswer: String;
    userType: String;
    personalId: String;
    approvedBy: String;
    profession: String;

}